Move asserts before usage of a variable
Before this commit, there were two asserts that were checked
only after both of the variables were used. This means that instead of
asserting, Krita would crash. It probaby doesn't change much in case of
Q_ASSERTs (it would be different with asserts that allows recovering)
but nonetheless, it makes no sense to assert after using a variable.
This commit moves the asserts above the usage.
BUG:421076