Fix autosave suggesting the autosave-formatted filename
Before this commit, if you recover an autosave for a saved file
(for example after a crash), Krita would open the autosave similarly
to how it opens other files: it would assume that the filename
that was opened is the filename to save to.
It did ask for confirmation, but beginner users might've saved it still
in the autosave-formatted filename.
This commit changes it so the autosaved file gets the filename
of the original file.
Krita still asks for confirmation even when clicked Save,
and even asks if it can overwrite the original file.

BUG:409950