Change autosave recovery dialog for existing files
Before this commit, the autosave recovery dialog was confusing
for our users. This commit makes the message hopefully more clear.
The user can choose which file to open (and is provided with thumbnails
so they are more informed what both of the files contain)
and there is a warning that discarding the autosave file means
losing it.

Warning: this commit contains new strings to translate.

CCMAIL:kde-i18n-doc@kde.org