Fix Histogram updates after opening a flat file
Before this commit, histogram would not update if you open a file
with a single layer and operate only on layers other than the original
one.

It was because the histogram widget would save the paint device
of the projection when the view was created. Projection paint device can
change in various cases: for example in this one, the paint device for
the projection was the reused projection of the first layer. When
creating a new layer, there was a new paint device created for the
projection, so the paint device

This commit fixes it by always adding the canvas as a parameter
when updating the histogram.

BUG:420014
(cherry picked from commit 243a413778e54de14225172e55bf81b28047fdc8)