Fixed canvas "jump" on finger touches.
Further touch handling fixes!

In https://invent.kde.org/graphics/krita/-/merge_requests/333#note_47474
@dkazakov observed that the canvas "jumps" randomly when the user's
fingers leave or touch the canvas.

This is due to the QTouchEvent with unreliable touch positions also
firing in the form of a QGestureEvent with equally dubious positions,
and that event bubbling up all the way to the parent
QAbstractScrollArea.

So we just prevent QGestureEvent events from bubbling up to the parent
QAbstractScrollArea, and call it a day.

BUG:421295