Fix the way URIs are handled
Prior to this commit, URIs were expected to resolve to
some file on Android, but this assumption turned out to
be flawed. URIs may or may not resolve. What is expected
way of opening a file using them is to either use InputStreams
or File Descriptors. Later is being used out of the two.

BUG:422058

# Conflicts:
#	libs/ui/KisMainWindow.cpp