Created a shader for drawing brush tool outline.
Made a new GLSL shader for drawing the brush tool's preview outline.
To do so, an adjustment to Krita's render pipeline was made:
On platforms where render to texture (FBO) is supported, the canvas is
now rendered to an intermediate texture which is used as a uniform input
in the outline shader (and possibly other overlays in the future).

CCBUG:415772
Coded with Eoin O'Neill.