Fix arrow markers to look better
The markers were generated in Inkscape. And there they had a stroke of
1pt. But(!) in Inkscape, the meaning of pt/px is a bit different. In
Inkscape 1px equals 1mm. But is Krita 1px is more or less equal to 1 pt.
Therefore the arrows looked ugly.

This patch just sets the arrow stroke width to a fixed value in 'px',
which solves the uncertainty caused by using configurable values.
2.8 = 72 pt/in / 25.4mm/in

BUG:402250