Remove old workaround for popup palette hiding
Commit 591684144731c0e8516bfd065fa812ff3e5f9c00 and
Commit ae734675be9f92173a9c44516a205edefb19ee97
added this to prevent picking up the new foreground color as
used by closing the palette with a left-click on the canvas.

The paint tools however have been refactored since then and
the "TODO" is no longer present in kis_tool_freehand.cc,
so remove the remains of this workaround.

BUG:415106