Krita 5: New Animation Workspace.
We've also taken this opportunity to make changes to Krita's Animation
Workspace.

Docked to the top is the standard Tool Box.

Docked to the upper-left are the Palette Docker, Tool Options Docker &
Overview Docker.

Docked to the lower-left are the Layers Docker and Compositions Docker.

Docked to the bottom are the Animation Timeline Docker, Animation Curves
Docker, and Onion Skins Docker.