Fix "Stroke Selection" when any selection tool is activated
This patch is a bit hackish. The actual bug is caused by the per-tool
opacity patch (6daf2cb7), which is still planned to be refactored.

The problem is that we store opacity and the preset(!), not in the
resource server itself. Therefore, if any paint tool is first activated
when there is no preset available, then it will remember default opacity
(which was 0.0 before this patch) and all non-painting tools will use this
opacity as default.

In the future we should refactor per-tool opacity, so that it would not
write to the presets (which makes them dirty when switching tools, which
is a bug). The non-painting tools should have some flag, that would tell
the resource provider that the opacity should be stored not in the preset,
but separately.

BUG:421752