Lens blur: Map between translated and untranslated strings in filter config
The filter stores the shape/sides parameter as a string taken directly
from the ui, instead of an enum or an index. This patch maps the
translated string back to the English string so we store the config
in English, and show the value translated.

BUG:422999
(cherry picked from commit 6ba7618cadeed70f7367711e840fda7edc1801a3)