Fix resizing KisVisualColorSelector in single slider configuration
Get rid of the mostly unused layout and handle geometry manually.
It'd probably take more code to define all hints for automatic layout,
also it might be easier to implement a HDR capable version if
the selector shapes are not actual widgets at all.

Finally, add a minimum size hint to KisVisualColorSelector