Add automatic parsing and error reporting
Thanks Wolthera van Hövell for the idea.
Expressions are now automagically parsed (after signal compression) and
errors are reported into a list to the effect.
The list is contained within the editor provided by SeExpr, see
https://invent.kde.org/lsegovia/seexpr/commit/75c41b41ea7f24d0710cf359c1fb5c2d3467fadc
As an addition of my own, I fixed number highlighting and added an
underline to each error's location.