Fix contiguous selection tool working on projection
Before this commit, the tool tried to access paint device of the layer
and return if there was none (which is the case for group layers).
This commit changes it to look for projection() instead, as it was
before the color labeled layers patch.

Using the paind device seemed to cause two bugs. It wasn't working on
group layers and clicking outside of the canvas didn't deselect.

BUG:423336
BUG:423237