Remove checkboxes in bundle manager
Before this commit, there were checkboxes in the bundle manager
that were supposed to show whether the bundle is active or not.
Because it's  quite difficult to make the checkboxes work the way
the user would expect them to work, I decided to switch it
to a bit different approach with borders around bundles that are
active. This commit adds borders and removes checkboxes.