Animation: Prev/Next Matching Keyframe Actions.
We just added a couple more actions for selecting the previous and next
keyframe with the same color label as the currently active keyframe.

This allows animators to organize various keyframes in their sequence
by color label, and then easily jump between keyframes with matching
color labels applied.

As of now, there is no GUI button or default shortcut for this action,
so the user will need to set one.

This helpful new feature was inspired by a request from Krita user, Ahmet Tabak.
Thanks for the contribution Ahmet!