Compositions Docker now Allows for the Exporting of Animations
Other changes needed to be made to get this to happen: Specifically,
the plugin `extensions/animationrenderer` is now removed and built
internally into Krita so that shared access is possible.

Compositions Docker's export settings are much simpler than the
Animation Export window for the time being. If users need more
control of export settings, we can come back to this. For the time
being, we simply use the previously used export settings for animation.