Tweak KisVisualEllipticalSelectorShape::getSpaceForSquare()
Give it a pixel more and make sure bottom right margin is
symmetric to top left.

Support KisVisualEllipticalSelectorShape with offset to parent

It used to assume top-left corner is (0, 0) when calculating the bounds
for inner controls.
Also clear some confusion with QRect's strange interpretation of bottomRight(),
and warn if the given rect does not contain the widget's geometry.