Finalize bundling support of SeExpr scripts
This adds the missing functionality to create, save and bundle SeExpr
scripts.

It includes:

* dialog to save and overwrite presets
* KisSharedPtr definition, to automate memory management inside the
dialogs
* slots and signals to communicate selected preset to SeExpr
* integration of SeExpr resources with the bundle manager

Everything is hidden behind the HAVE_SEEXPR flag, which is added if
CMake detects the presence of the SeExpr library.

I have also adjusted the path of the SeExpr scripts folder, as it did
not match the expectations of the bundle manager (Ko/Kis + resType,
where resType is the folder name in .local/share/krita).

Thanks to Scott Petrovic and Anna Medonosová for their work on paint ops
and gamut masks. Most of the stuff here is adjustments to match SeExpr's
requirements.