Fix color conversion from SeExpr to KisPaintDevice
According to Wolthera, in Qt versions that do not have scRGB (< 5.12??),
the current version of my code will make QColor complain if the values
are not clamped.

This version attempts to detect if the underlying color profile supports
extended range RGB, and if so, send the colors directly. Otherwise, they
are clamped (integer RGB) or converted AND clamped (non-RGB spaces).
The latter could be lifted with a manual conversion from a floating
point KoColorProfile, but our code does not provide a default for RGB.