Remove canvas code meant for handling the old text shapes
And incidentally, now we probably know why the old text shape
sometimes rendered in a broken way, broken enough that we removed
it: it was rendered once again on top of the canvas _as_ well as
in the projection of the vector layer. If the alignment wasn't
perfect, you'd get those ghosted double-printed characters that
made the old textshape unusable.

We could't kept it after all, I fear...

CCMAIL:kimageshop@kde.org