Add a workaround for too big brushes generated by Masking Brush option
The masking brush is scaled by the "master" brush size value. That is,
it may become much larger than the user-defined limit.

This patch is a workaround that limits the size of the masking brush
by 3*brughLimit size. The patch doesn't notify the user about the cropping,
neither adjusts it the sliders limits, hence "workaround".

TODO:

1) Implement a warning notifying the user that his masking
   brush has been cropped

2) Make sure that the sliders in KisMaskingBrushOption have
   correct limits (right now they are limited by usual
   maximumBrushSize)

Thanks Defresne Thierry for tracking down the bug and providing
a draft patch! :)

BUG:423572