Fix saving broken .ora file in case of empty layer
Before this commit, saving an empty layer to .ora would result in a
broken file because the result PNG for the layer would be 0x0.
This commit artificially increases the size of the layer with one pixel
when the layer is empty to make sure there is some data (even if
invisible) to save to PNG file.
BUG:423741


(cherry picked from commit 925bf97f8d4ed1101267fd86d99536069a05f969)