Don't save x and y of groups in .ora files
Before this commit Krita would save position attributes for
group layers. However Krita and Gimp's handling of those attributes
differs from the behaviour MyPaint has. For the layer inside the group,
the position attributes are considered relative to the group layer
position (they are summed). In Krita and Gimp, the position of the layer
inside the group is considered absolute.

For example this layer structure:
> group layer x=100 y=100
>> paint layer x=100 y=100

In Krita and Gimp, the paint layer will appear in x=100 y=100
on the canvas, while in MyPaint it will appear in x=200 y=200.

This commit always saves group layers with x=0 and y=0 to avoid this
issue.

CCBUG:423088