Don't save fake nodes in .ora files
Before this commit, Krita would try to save reference images layer into
.ora files. Since reference images layer doesn't have any projection
(and the name for the layer is 'No Name'...), it would result in a ghost
new layer on top of the layer stack called 'No Name'.
Since there is no projection nor other information, there is no point
in saving it to .ora files.
This commit omits reference images layer (and other fake nodes)
when saving to .ora.

CCBUG:423741