Get the image's dimensions from the KisPaintDevice
Before this commit, the generator will render the whole texture in each
tile. While this is correct right now (the KisGeneratorLayer job updates
as one whole tile), once it is multithreaded, it will render N copies of
the texture, one inside each tile.

This commit gets the real bounds from the KisPaintDevice and adjusts the
stride, width, and height accordingly.