Reimplemented ability to add opacity keyframes to the Curves docker.
Removing AnimationDocker caused a regression where the curves docker
could no longer have opacity keys added. For the time being, the ability
to add opacity keys has now been restored by having the add key button
work as expected.

**How it should work once more than one key is supported**

When we eventually add Transform Mask keys, the user should be able to
drop down a series of check boxes for the "add key" button which gives
them the ability to enable and disable certain channels. This should be
managed by adding actions to a QToolButton that enable and disable
all supported channels. Since opacity is really the only channel we
currently support, this isn't necessary, but we should keep this in mind.