Initial commit of the Multigrid generator.
This is a generator that implements de Bruijn's 1981 algorithm
to generate Penrose tesselations. The algorithm does this at
dimension 5 with offsets .2 and .4. At all the other values the
result isn't an official Penrose tesselation, but still looks cool.

These tesselations show up in nature as Quasicrystals.

There's still a number of extensions and papercuts to be adressed.