Fix crash on curve adjustment filter in CMYK document
Before this commit, Krita will crash on accessing Saturation
or other "fake" channel in the curve adjustment filter dialog.
This commit fixes it.
The problem was in recent change in the PerChannelFilter,
it would access the actual channels of the color space
to set min and max of the dialog. Since Saturation isn't
a channel of the CMYK color space, it would crash.

I'm not 100% sure the code here doesn't have any more
issues but I don't know how it exactly should be written.
In any case, this commit improves the situation by setting min
and max manually in case of "fake" channels. This is consistent
with the old behaviour before dffa076018d6bf98ec0.

BUG:424187
(cherry picked from commit cc273e4985c42f940c60908be6d02ed47554652e)