Fix no virtual destrutor in KisResourcesInterfacePrivate
KisLocalStrokeResourcesPrivate class which inherits after
KisResourcesInterfacePrivate has a compiler-generated destructor
(because of QList<KoResourceSP>). This class is being upcasted
to KisResourcesInterfacePrivate in th constructor of
KisLocalStrokeResources class. And then it's being deleted
at the end of the lifetime of this class becuase it's encapsulated
in the QScopedPointer.

Hence it's possible that it was KisLocalStrokeResourcesPrivate
that has a destructor but the destructor wasn't called
(because the destructor of the KisResourceInterfacePrivate class
was called which is not virtual).

This commit adds a custom virtual destructor.

CID 304984