Fix saving custom brush presets
Before this commit, brush presets were saved without the extension.
In the old code, before the resource rewrite, the extension was added
to the name submitted by the user and only later saved,
and this commit ensures it is being added in the resource rewrite
version of this code, too.