Fix gradients that depend on Fg/Bg colors
When a gradient is used for a stroke, it is cloned via
cloneAndBakeVariableColors(), which basically converts it into
a normal gradient with statci colors.

The current Fg/Bg colors are fetched via provided
KoCanvasResourcesInterfaceSP pointer.

The patch also implements a version of generatePreview() that
takes requiredCanvasResources() into account, but it is not used yet.