Initialize m_firstRedo in FlipFlopCommand
Before this commit one constructor of the FlipFlopCommand
initialized m_firstRedo to true, and the other constructor
didn't initialize it at all.
This commit removes the initialization of m_firstRedo in the
initialization list of the constructor and instead initializes it
in the place of definition.

Since constructors don't seem to be doing something different - both
take one parameter telling if the command is finalizing or not, and
a parent command - I assumed that both would want m_firstRedo to be
initialized to the same value, which is 'true'. However considering
that I believe most compilers would initialize the value to false,
this can change the behaviour, so this commit might not be.

CID 248600