Fix inability to save SeExpr presets
This does not fix another underlying bug when loading the preset
(it saves it on the bundle, and then loads it from the global
folder instead).
I'm told by @rempt that this one is fixed in another MR, so
I'll leave it as is.

CCMAIL: boud@valdyas.org
CCMAIL: dimula73@gmail.com