This fixes the bug where this brush slowly outputs everything to transparent.
The bug was caused by the fact that to sample the input pixels, the crossdevice color picker was used, which in turn got access to subpixel data by utilizing the randomsubaccessor, which was getting that data by mixing neighbouring pixels via the KoColorMixOp. However, the randomsubaccessor isn't able to generate perfect sums-to-255 weights every single time, meaning transparency was added.

This commit fixes that by passing a sum of weights to the KoColorMixOp that only defaults to 255.

Special thanks to Dmitry for figuring this out.

Note: there's still an issue with colors mixing darker in 8bit, we're still trying to figure that one out. But at the least there's no transparency.

CCBUG:290383