Fix darkening of colors using the deform brush.
Based of dmitry's initial patch, this adds a hack to ensure
the appropriate division function is selected. We should be using
KoColorSpaceMaths and templates for this, but it is not clear to me
how this should be done, and I even suspect it requires a bigger refactor.

This however is fast enough and makes the brush work as expected in 8bit
integer and 32bit float versions of all the available color models.

BUG:290383