Update SeExpr version
Before this commit, Krita would crash when creating vector variables.

When initializing vector variables' widgets with a specified range,
or attempting to change their value, SeExpr would try to access an
uninitialized member. The check for null does not work because it was
never initialized in the first place.

Additionally, I fixed this widget so that it would correctly allow
vectors with explicit [0, 1] range to be previewed as color variables.