Fix update of file layers created via "Convert to" menu
`FileSystemWatcherWrapper` works with absolute paths, but
`KisSafeDocumentLoader` works with the original path as it
was passed to the layer. We should unify these paths in
`KisSafeDocumentLoader::fileChanged()`.

The patch also fixes lost connections, when the file is
temporary deleted (moved) and its recreation is delayed
by something. Previously, FileSystemWatcherWrapper::slotFileChanged
could just forget about it. Now it maintains a list of missing
files in `m_lostFilesAbsenceCounter` and tries to reconnect to
them with 200 ms intervals.

PS:
The tracked file may be removed/recreated because of QSaveFile
functionality.

BUG:411790