Don't use document title when warning about unsaved changes
Before this commit, Krita would use document title
(which is usually the filename that was first used
for that document without the extension).
That would cause a confusion when "Save As" feature is used
since Krita would stll hold the old name as the document title.
This commit changes it to always use the filename instead
(in the unsaved changes warning dialog).

BUG:422638