Show display names in Color Management -> Display
Before this commit, user would have to figure out on their
own which display is which. Because of a bug in the system
assigning color profiles to displays, the automatic way doesn't
always work so the user needs to assign color profiles on their own.
Before this commit the dialog only showed "Screen 1, Screen 2".
This commit adds following information:
- name
- if model is not null, add model; if it is null, add manufacturer info
- resolution
It is also shortened so the maximum length is 25 characters.

CCBUG:412943