Fix first redo() after paint device color space conversion
There is a bug in all versions of Krita since November, 5th 2019.

BUG
===

1) Change image color space from RGBA to GrayA.
2) Undo with Ctrl+Z
3) Redo with Ctrl+Shift+Z

The image looks as if the color space has changed back to GrayA,
but it is not so. Only the color space of the projection has changed
to GrayA, the layers are still in RGBA. It can be seen by the state
of Advanced Color Selector (it shows colors instead of shades of gray)
and in layers' properties.

WORKAROUND
==========

1) Press Undo (Ctrl+Z) and Redo (Ctrl+Shift+Z) again.

The image will enter the correct state. All further undo/redo
operations for this particular action will be executed correctly.

TECHNICAL DETAILS
=================

The bug was called by commit e39e6379564b5f6c4c72960a6e1ee3ca7716d5bd,
which caused DeviceChangeColorSpaceCommand to skip the first redo of
the operation, which was not in its design.

The bug has been caught by KisColorizeMaskTest.

BUG:416584
CC:kimageshop@kde.org