Fix KisUpdateSchedulerTest
The test should have used KisTestableUpdaterContext, but due to some
old commit in 2015 it started to use a normal updater context. It
meant that the jobs were executed in parallel to the test being run.
That caused race conditions caught by CI.

The patch also simplifies KisUpdaterContext and removes all virtuality
from it (it is not QObject since lockfree tiles refactoring).