Fix a theoretical race condition in KisMementoManager::registerTileDeleted()
The crash could happen if one thread was changing the default pixel of
a paint device while ome other thread was still writing into it. Such
event should not be possible in real Krita. Firstly, it breaks "single
writer for a pixel" invariant. Secondly, the result of such write is
undefined.

Though it would be nice if the data manager would not crash at least in
such circumstances.

The patch fixes KisTiledDataManagerTest::stressTest(), that caught this
issue.